Extract CLI groups to modules
