+++ perun/cli.py
@@ -42,30 +42,27 @@ the flexibility of Perun's usage.
 import os
 import pkgutil
 import sys
-import distutils.util as dutils
-import termcolor
 
 import click
 
 import perun.collect
-import perun.check.factory as check
 import perun.fuzz.factory as fuzz
 import perun.logic.commands as commands
-import perun.logic.runner as runner
 import perun.logic.pcs as pcs
 import perun.logic.config as perun_config
-import perun.logic.temp as temp
-import perun.logic.stats as stats
 import perun.postprocess
 import perun.profile.helpers as profiles
 import perun.utils as utils
-import perun.utils.script_helpers as scripts
 import perun.utils.cli_helpers as cli_helpers
 import perun.utils.log as perun_log
 import perun.view
 from perun.utils.exceptions import UnsupportedModuleException, UnsupportedModuleFunctionException, \
     NotPerunRepositoryException, IncorrectProfileFormatException, EntryNotFoundException, \
     MissingConfigSectionException, ExternalEditorErrorException
+import perun.cli_groups.check_cli as check_cli
+import perun.cli_groups.config_cli as config_cli
+import perun.cli_groups.run_cli as run_cli
+import perun.cli_groups.utils_cli as utils_cli
 
 
 __author__ = 'Tomas Fiedor'
@@ -112,176 +109,6 @@ def cli(verbose=0, no_pager=False, **_):
     # set the verbosity level of the log
     if perun_log.VERBOSITY < verbose:
         perun_log.VERBOSITY = verbose
-
-
-@cli.group()
-@click.option('--local', '-l', 'store_type', flag_value='local',
-              help='Sets the local config, i.e. ``.perun/local.yml``, as the source config.')
-@click.option('--shared', '-h', 'store_type', flag_value='shared',
-              help='Sets the shared config, i.e. ``shared.yml.``, as the source config')
-@click.option('--nearest', '-n', 'store_type', flag_value='recursive', default=True,
-              help='Sets the nearest suitable config as the source config. The'
-              ' lookup strategy can differ for ``set`` and '
-              '``get``/``edit``.')
-@click.pass_context
-def config(ctx, **kwargs):
-    """Manages the stored local and shared configuration.
-
-    Perun supports two external configurations:
-
-        1. ``local.yml``: the local configuration stored in ``.perun``
-           directory, containing the keys such as specification of wrapped
-           repository or job matrix used for quick generation of profiles (run
-           ``perun run matrix --help`` or refer to :doc:`jobs` for information
-           how to construct the job matrix).
-
-        2. ``shared.yml``:  the global configuration shared by all perun
-           instances, containing shared keys, such as text editor, formatting
-           string, etc.
-
-    The syntax of the ``<key>`` in most operations consists of section
-    separated by dots, e.g. ``vcs.type`` specifies ``type`` key in ``vcs``
-    section. The lookup of the ``<key>`` can be performed in three modes,
-    ``--local``, ``--shared`` and ``--nearest``, locating or setting the
-    ``<key>`` in local, shared or nearest configuration respectively (e.g. when
-    one is trying to get some key, there may be nested perun instances that do
-    not contain the given key). By default, perun operates in the nearest
-    config mode.
-
-    Refer to :doc:`config` for full description of configurations and
-    :ref:`config-types` for full list of configuration options.
-
-    E.g. using the following one can retrieve the type of the nearest perun
-    instance wrapper:
-
-    .. code-block:: bash
-
-        $ perun config get vcs.type
-        vcs.type: git
-    """
-    ctx.obj = kwargs
-
-
-@config.command('get')
-@click.argument('key', required=True, metavar='<key>',
-                callback=cli_helpers.config_key_validation_callback)
-@click.pass_context
-def config_get(ctx, key):
-    """Looks up the given ``<key>`` within the configuration hierarchy and returns
-    the stored value.
-
-    The syntax of the ``<key>`` consists of section separated by dots, e.g.
-    ``vcs.type`` specifies ``type`` key in ``vcs`` section. The lookup of the
-    ``<key>`` can be performed in three modes, ``--local``, ``--shared`` and
-    ``--nearest``, locating the ``<key>`` in local, shared or nearest
-    configuration respectively (e.g. when one is trying to get some key, there
-    may be nested perun instances that do not contain the given key). By
-    default, perun operates in the nearest config mode.
-
-    Refer to :doc:`config` for full description of configurations and
-    :ref:`config-types` for full list of configuration options.
-
-    E.g. using the following can retrieve the type of the nearest perun
-    wrapper:
-
-    .. code-block:: bash
-
-        $ perun config get vcs.type
-        vcs.type: git
-
-        $ perun config --shared get general.editor
-        general.editor: vim
-    """
-    try:
-        commands.config_get(ctx.obj['store_type'], key)
-    except MissingConfigSectionException as mcs_err:
-        perun_log.error("error while getting key '{}': {}".format(
-            key, str(mcs_err)
-        ))
-
-
-@config.command('set')
-@click.argument('key', required=True, metavar='<key>',
-                callback=cli_helpers.config_key_validation_callback)
-@click.argument('value', required=True, metavar='<value>')
-@click.pass_context
-def config_set(ctx, key, value):
-    """Sets the value of the ``<key>`` to the given ``<value>`` in the target
-    configuration file.
-
-    The syntax of the ``<key>`` corresponds of section separated by dots, e.g.
-    ``vcs.type`` specifies ``type`` key in ``vcs`` section. Perun sets the
-    ``<key>`` in three modes, ``--local``, ``--shared`` and ``--nearest``,
-    which sets the ``<key>`` in local, shared or nearest configuration
-    respectively (e.g.  when one is trying to get some key, there may be nested
-    perun instances that do not contain the given key). By default, perun will
-    operate in the nearest config mode.
-
-    The ``<value>`` is arbitrary depending on the key.
-
-    Refer to :doc:`config` for full description of configurations and
-    :ref:`config-types` for full list of configuration options and their
-    values.
-
-    E.g. using the following can set the log format for nearest perun instance
-    wrapper:
-
-    .. code-block:: bash
-
-        $ perun config set format.shortlog "| %source% | %collector% |"
-        format.shortlog: | %source% | %collector% |
-    """
-    commands.config_set(ctx.obj['store_type'], key, value)
-
-
-@config.command('edit')
-@click.pass_context
-def config_edit(ctx):
-    """Edits the configuration file in the external editor.
-
-    The used editor is specified by the :ckey:`general.editor` option,
-    specified in the nearest perun configuration..
-
-    Refer to :doc:`config` for full description of configurations and
-    :ref:`config-types` for full list of configuration options.
-    """
-    try:
-        commands.config_edit(ctx.obj['store_type'])
-    except (ExternalEditorErrorException, MissingConfigSectionException) as editor_exception:
-        perun_log.error("could not invoke external editor: {}".format(str(editor_exception)))
-
-
-@config.command('reset')
-@click.argument('config_template', required=False, default='master',
-                metavar='<template>')
-@click.pass_context
-def config_reset(ctx, config_template):
-    """Resets the configuration file to a sane default.
-
-    If we are resetting the local configuration file we can specify a <template> that
-    will be used to generate a predefined set of options. Currently we support the following:
-
-      1. **user** configuration is meant for beginner users, that have no experience with Perun and
-      have not read the documentation thoroughly. This contains a basic preconfiguration that should
-      be applicable for most of the projects---data are collected by :ref:`collectors-time` and are
-      automatically registered in the Perun after successful run. The performance is checked using
-      the :ref:`degradation-method-aat`. Missing profiling info will be looked up automatically.
-
-      2. **developer** configuration is meant for advanced users, that have some understanding of
-      profiling and/or Perun. Fair amount of options are up to the user, such as the collection of
-      the data and the commands that will be profiled.
-
-      3. **master** configuration is meant for experienced users. The configuration will be mostly
-      empty.
-
-    See :ref:`config-templates` to learn more about predefined configuration options.
-    """
-    try:
-        commands.config_reset(ctx.obj['store_type'], config_template)
-    except NotPerunRepositoryException as npre:
-        perun_log.error("could not reset the {} configuration: {}".format(
-            ctx.obj['store_type'], str(npre)
-        ))
 
 
 def configure_local_perun(perun_path):
@@ -769,148 +596,6 @@ def collect(ctx, **kwargs):
     ctx.obj = kwargs
 
 
-@cli.group()
-@click.option('--output-filename-template', '-ot', default=None,
-              callback=cli_helpers.set_config_option_from_flag(
-                  perun_config.runtime, 'format.output_profile_template', str
-              ), help='Specifies the template for automatic generation of output filename'
-              ' This way the file with collected data will have a resulting filename w.r.t '
-              ' to this parameter. Refer to :ckey:`format.output_profile_template` for more'
-              ' details about the format of the template.')
-@click.option('--minor-version', '-m', 'minor_version_list', nargs=1, multiple=True,
-              callback=cli_helpers.minor_version_list_callback, default=['HEAD'],
-              help='Specifies the head minor version, for which the profiles will be collected.')
-@click.option('--crawl-parents', '-c', is_flag=True, default=False, is_eager=True,
-              help='If set to true, then for each specified minor versions, profiles for parents'
-                   ' will be collected as well')
-@click.option('--force-dirty', '-f', is_flag=True, default=False,
-              callback=cli_helpers.unsupported_option_callback,
-              help='If set to true, then even if the repository is dirty, '
-                   'the changes will not be stashed')
-@click.pass_context
-def run(ctx, **kwargs):
-    """Generates batch of profiles w.r.t. specification of list of jobs.
-
-    Either runs the job matrix stored in local.yml configuration or lets the
-    user construct the job run using the set of parameters.
-    """
-    ctx.obj = kwargs
-
-
-@run.command()
-@click.pass_context
-@click.option('--without-vcs-history', '-q', 'quiet', is_flag=True, default=False,
-              help="Will not print the VCS history tree during the collection of the data.")
-def matrix(ctx, quiet, **kwargs):
-    """Runs the jobs matrix specified in the local.yml configuration.
-
-    This commands loads the jobs configuration from local configuration, builds
-    the `job matrix` and subsequently runs the jobs collecting list of
-    profiles. Each profile is then stored in ``.perun/jobs`` directory and
-    moreover is annotated using by setting :preg:`origin` key to current
-    ``HEAD``. This serves as check to not assing such profiles to different
-    minor versions.
-
-    The job matrix is defined in the yaml format and consists of specification
-    of binaries with corresponding arguments, workloads, supported collectors
-    of profiling data and postprocessors that alter the collected profiles.
-
-    Refer to :doc:`jobs` and :ref:`jobs-matrix` for more details how to specify
-    the job matrix inside local configuration and to :doc:`config` how to work
-    with Perun's configuration files.
-    """
-    kwargs.update({'minor_version_list': ctx.obj['minor_version_list']})
-    kwargs.update({'with_history': not quiet})
-    if runner.run_matrix_job(**kwargs) != runner.CollectStatus.OK:
-        perun_log.error("job specification failed in one of the phases")
-
-
-@run.command()
-@click.option('--cmd', '-b', nargs=1, required=True, multiple=True,
-              help='Command that is being profiled. Either corresponds to some'
-              ' script, binary or command, e.g. ``./mybin`` or ``perun``.')
-@click.option('--args', '-a', nargs=1, required=False, multiple=True,
-              help='Additional parameters for <cmd>. E.g. ``status`` or '
-              '``-al`` is command parameter.')
-@click.option('--workload', '-w', nargs=1, required=False, multiple=True, default=[''],
-              help='Inputs for <cmd>. E.g. ``./subdir`` is possible workload'
-              'for ``ls`` command.')
-@click.option('--collector', '-c', nargs=1, required=True, multiple=True,
-              type=click.Choice(utils.get_supported_module_names('collect')),
-              help='Profiler used for collection of profiling data for the'
-              ' given <cmd>')
-@click.option('--collector-params', '-cp', nargs=2, required=False, multiple=True,
-              callback=cli_helpers.yaml_param_callback,
-              help='Additional parameters for the <collector> read from the'
-              ' file in YAML format')
-@click.option('--postprocessor', '-p', nargs=1, required=False, multiple=True,
-              type=click.Choice(utils.get_supported_module_names('postprocess')),
-              help='After each collection of data will run <postprocessor> to '
-              'postprocess the collected resources.')
-@click.option('--postprocessor-params', '-pp', nargs=2, required=False, multiple=True,
-              callback=cli_helpers.yaml_param_callback,
-              help='Additional parameters for the <postprocessor> read from the'
-              ' file in YAML format')
-@click.pass_context
-def job(ctx, **kwargs):
-    """Run specified batch of perun jobs to generate profiles.
-
-    This command correspond to running one isolated batch of profiling jobs,
-    outside of regular profilings. Run ``perun run matrix``, after specifying
-    job matrix in local configuration to automate regular profilings of your
-    project. After the batch is generated, each profile is taged with
-    :preg:`origin` set to current ``HEAD``. This serves as check to not assing
-    such profiles to different minor versions.
-
-    By default the profiles computed by this batch job are stored inside the
-    ``.perun/jobs/`` directory as a files in form of::
-
-        bin-collector-workload-timestamp.perf
-
-    In order to store generated profiles run the following, with ``i@p``
-    corresponding to `pending tag`, which can be obtained by running ``perun
-    status``::
-
-        perun add i@p
-
-    .. code-block:: bash
-
-        perun run job -c time -b ./mybin -w file.in -w file2.in -p normalizer
-
-    This command profiles two commands ``./mybin file.in`` and ``./mybin
-    file2.in`` and collects the profiling data using the
-    :ref:`collectors-time`. The profiles are afterwards normalized with the
-    :ref:`postprocessors-normalizer`.
-
-    .. code-block:: bash
-
-        perun run job -c complexity -b ./mybin -w sll.cpp -cp complexity targetdir=./src
-
-    This commands runs one job './mybin sll.cpp' using the
-    :ref:`collectors-trace`, which uses custom binaries targeted at
-    ``./src`` directory.
-
-    .. code-block:: bash
-
-        perun run job -c mcollect -b ./mybin -b ./otherbin -w input.txt -p normalizer -p clusterizer
-
-    This commands runs two jobs ``./mybin input.txt`` and ``./otherbin
-    input.txt`` and collects the profiles using the :ref:`collectors-memory`.
-    The profiles are afterwards postprocessed, first using the
-    :ref:`postprocessors-normalizer` and then with
-    :ref:`postprocessors-regression-analysis`.
-
-    Refer to :doc:`jobs` and :doc:`profile` for more details about automation
-    and lifetimes of profiles. For list of available collectors and
-    postprocessors refer to :ref:`collectors-list` and
-    :ref:`postprocessors-list` respectively.
-    """
-    kwargs.update({'minor_version_list': ctx.obj['minor_version_list']})
-    kwargs.update({'with_history': True})
-    if runner.run_single_job(**kwargs) != runner.CollectStatus.OK:
-        perun_log.error("job specification failed in one of the phases")
-
-
 @cli.command('fuzz')
 @click.option('--cmd', '-b', nargs=1, required=True,
               help='The command which will be fuzzed.')
@@ -939,395 +624,6 @@ def job(ctx, **kwargs):
 def fuzz_cmd(**kwargs):
     """Performs fuzzing for the specified command according to the initial sample of workload."""
     fuzz.run_fuzzing_for_command(**kwargs)
-
-
-@cli.group('check')
-@click.option('--compute-missing', '-c',
-              callback=cli_helpers.set_config_option_from_flag(
-                  perun_config.runtime, 'degradation.collect_before_check'),
-              is_flag=True, default=False,
-              help='whenever there are missing profiles in the given point of history'
-              ' the matrix will be rerun and new generated profiles assigned.')
-@click.option('--models-type', '-m', nargs=1, required=False, multiple=False,
-              type=click.Choice(check.get_supported_detection_models_strategies()),
-              default=check.get_supported_detection_models_strategies()[0],
-              help="The detection models strategies predict the way of executing "
-                   "the detection between two profiles, respectively between relevant "
-                   "kinds of its models. Available only in the following detection "
-                   "methods: Integral Comparison (IC) and Local Statistics (LS).")
-def check_group(**_):
-    """Applies for the points of version history checks for possible performance changes.
-
-    This command group either runs the checks for one point of history (``perun check head``) or for
-    the whole history (``perun check all``). For each minor version (called the `target`) we iterate
-    over all of the registered profiles and try to find a predecessor minor version (called the
-    `baseline`) with profile of the same configuration (by configuration we mean the tuple of
-    collector, postprocessors, command, arguments and workloads) and run the checks according to the
-    rules set in the configurations.
-
-    The rules are specified as an ordered list in the configuration by
-    :ckey:`degradation.strategies`, where the keys correspond to the configuration (or the type) and
-    key `method` specifies the actual method used for checking for performance changes. The applied
-    methods can then be either specified by the full name or by its short string consisting of all
-    first letter of the function name.
-
-    The example of configuration snippet that sets rules and strategies for one project can be as
-    follows:
-
-      .. code-block:: yaml
-
-      \b
-          degradation:
-            apply: first
-            strategies:
-              - type: mixed
-                postprocessor: regression_analysis
-                method: bmoe
-              - cmd: mybin
-                type: memory
-                method: bmoe
-              - method: aat
-
-    Currently we support the following methods:
-
-      \b
-      1. Best Model Order Equality (BMOE)
-      2. Average Amount Threshold (AAT)
-    """
-    should_precollect = dutils.strtobool(str(
-        perun_config.lookup_key_recursively('degradation.collect_before_check', 'false')
-    ))
-    precollect_to_log = dutils.strtobool(str(
-        perun_config.lookup_key_recursively('degradation.log_collect', 'false')
-    ))
-    if should_precollect:
-        print("{} is set to {}. ".format(
-            termcolor.colored('degradation.collect_before_check', 'white', attrs=['bold']),
-            termcolor.colored('true', 'green', attrs=['bold'])
-        ), end='')
-        print("Missing profiles will be freshly collected with respect to the ", end='')
-        print("nearest job matrix (run `perun config edit` to modify the underlying job matrix).")
-        if precollect_to_log:
-            print("The progress of the pre-collect phase will be stored in logs at {}.".format(
-                termcolor.colored(pcs.get_log_directory(), 'white', attrs=['bold'])
-            ))
-        else:
-            print("The progress of the pre-collect phase will be redirected to {}.".format(
-                termcolor.colored('black hole', 'white', attrs=['bold'])
-            ))
-
-
-@check_group.command('head')
-@click.argument('head_minor', required=False, metavar='<hash>', nargs=1,
-                callback=cli_helpers.lookup_minor_version_callback, default='HEAD')
-def check_head(head_minor='HEAD'):
-    """Checks for changes in performance between between specified minor version (or current `head`)
-    and its predecessor minor versions.
-
-    The command iterates over all of the registered profiles of the specified `minor version`
-    (`target`; e.g. the `head`), and tries to find the nearest predecessor minor version
-    (`baseline`), where the profile with the same configuration as the tested target profile exists.
-    When it finds such a pair, it runs the check according to the strategies set in the
-    configuration (see :ref:`degradation-config` or :doc:`config`).
-
-    By default the ``hash`` corresponds to the `head` of the current project.
-    """
-    print("")
-    check.degradation_in_minor(head_minor)
-
-
-@check_group.command('all')
-@click.argument('minor_head', required=False, metavar='<hash>', nargs=1,
-                callback=cli_helpers.lookup_minor_version_callback, default='HEAD')
-def check_all(minor_head='HEAD'):
-    """Checks for changes in performance for the specified interval of version history.
-
-    The commands crawls through the whole history of project versions starting from the specified
-    ``<hash>`` and for all of the registered profiles (corresponding to some `target` minor version)
-    tries to find a suitable predecessor profile (corresponding to some `baseline` minor version)
-    and runs the performance check according to the set of strategies set in the configuration
-    (see :ref:`degradation-config` or :doc:`config`).
-    """
-    print("[!] Running the degradation checks on the whole VCS history. This might take a while!\n")
-    check.degradation_in_history(minor_head)
-
-
-@check_group.command('profiles')
-@click.argument('baseline_profile', required=True, metavar='<baseline>', nargs=1,
-                callback=cli_helpers.lookup_any_profile_callback)
-@click.argument('target_profile', required=True, metavar='<target>', nargs=1,
-                callback=cli_helpers.lookup_any_profile_callback)
-@click.option('--minor', '-m', nargs=1, default=None, is_eager=True,
-              callback=cli_helpers.lookup_minor_version_callback, metavar='<hash>',
-              help='Will check the index of different minor version <hash>'
-                   ' during the profile lookup.')
-@click.pass_context
-def check_profiles(ctx, baseline_profile, target_profile, minor, **_):
-    """Checks for changes in performance between two profiles.
-
-    The commands checks for the changes between two isolate profiles, that can be stored in pending
-    profiles, registered in index, or be simply stored in filesystem. Then for the pair of profiles
-    <baseline> and <target> the command runs the performance chekc according to the set of
-    strategies set in the configuration (see :ref:`degradation-config` or :doc:`config`).
-
-    <baseline> and <target> profiles will be looked up in the following steps:
-
-        1. If profile is in form ``i@i`` (i.e, an `index tag`), then `ith`
-           record registered in the minor version <hash> index will be used.
-
-        2. If profile is in form ``i@p`` (i.e., an `pending tag`), then
-           `ith` profile stored in ``.perun/jobs`` will be used.
-
-        3. Profile is looked-up within the minor version <hash> index for a
-           match. In case the <profile> is registered there, it will be used.
-
-        4. Profile is looked-up within the ``.perun/jobs`` directory. In case
-           there is a match, the found profile will be used.
-
-        5. Otherwise, the directory is walked for any match. Each found match
-           is asked for confirmation by user.
-
-    """
-    print("")
-    check.degradation_between_files(
-        baseline_profile, target_profile, minor, ctx.parent.params['models_type']
-    )
-
-
-@cli.group('utils')
-def utils_group():
-    """Contains set of developer commands, wrappers over helper scripts and other functions that are
-    not the part of the main perun suite.
-    """
-    pass
-
-
-@utils_group.command()
-@click.argument('template_type', metavar='<template>', required=True,
-                type=click.Choice(['collect', 'postprocess', 'view', 'check']))
-@click.argument('unit_name', metavar='<unit>')
-@click.option('--no-before-phase', '-nb', default=False, is_flag=True,
-              help='If set to true, the unit will not have before() function defined.')
-@click.option('--no-after-phase', '-na', default=False, is_flag=True,
-              help='If set to true, the unit will not have after() function defined.')
-@click.option('--author', nargs=1,
-              help='Specifies the author of the unit')
-@click.option('--no-edit', '-ne', default=False, is_flag=True,
-              help='Will open the newly created files in the editor specified by '
-                   ':ckey:`general.editor` configuration key.')
-@click.option('--supported-type', '-st', 'supported_types', nargs=1, multiple=True,
-              help="Sets the supported types of the unit (i.e. profile types).")
-def create(template_type, **kwargs):
-    """According to the given <template> constructs a new modules in Perun for <unit>.
-
-    Currently this supports creating new modules for the tool suite (namely ``collect``,
-    ``postprocess``, ``view``) or new algorithms for checking degradation (check). The command uses
-    templates stored in `../perun/templates` directory and uses _jinja as a template handler. The
-    templates can be parametrized by the following by options (if not specified 'none' is used).
-
-    Unless ``--no-edit`` is set, after the successful creation of the files, an external editor,
-    which is specified by :ckey:`general.editor` configuration key.
-
-    .. _jinja: http://jinja2.pocoo.org/
-    """
-    try:
-        scripts.create_unit_from_template(template_type, **kwargs)
-    except ExternalEditorErrorException as editor_exception:
-        perun_log.error("while invoking external editor: {}".format(str(editor_exception)))
-
-
-@utils_group.group('temp')
-def temp_group():
-    """Provides a set of operations for maintaining the temporary directory (.perun/tmp/) of perun.
-    """
-    pass
-
-
-@temp_group.command('list')
-@click.argument('root', type=click.Path(), required=False, default='.')
-@click.option('--no-color', '-c', flag_value=True, default=False,
-              help='Disable the output coloring, useful for storing the output to file etc.')
-@click.option('--no-total-size', '-t', flag_value=True, default=False,
-              help='Do not show the total size of all the temporary files combined.')
-@click.option('--no-file-size', '-f', flag_value=True, default=False,
-              help='Do not show the size of each temporary file.')
-@click.option('--no-protection-level', '-p', flag_value=True, default=False,
-              help='Do not show the protection level of the temporary files.')
-@click.option('--sort-by', '-s', type=click.Choice(temp.SORT_ATTR), default=temp.SORT_ATTR[0],
-              help='Sorts the temporary files on the output.')
-@click.option('--filter-protection', '-fp', type=click.Choice(temp.PROTECTION_LEVEL),
-              default=temp.PROTECTION_LEVEL[0],
-              help='List only temporary files with the given protection level.')
-def temp_list(root, **kwargs):
-    """Lists the temporary files of the '.perun/tmp/' directory. It is possible to list only
-    files in specific subdirectory by supplying the ROOT path.
-
-    The path can be either absolute or relative - the base of the relative path is the tmp/
-    directory.
-    """
-    commands.print_temp_files(root, **kwargs)
-
-
-@temp_group.command('delete')
-@click.argument('path', type=click.Path(), required=True)
-@click.option('--warn', '-w', flag_value=True, default=False,
-              help='Warn the user (and abort the deletion with no files deleted) if protected files'
-                   ' are present.')
-@click.option('--force', '-f', flag_value=True, default=False,
-              help='If set, protected files are deleted regardless of --warn value.')
-@click.option('--keep-directories', '-k', flag_value=True, default=False,
-              help='If path refers to directory, empty tmp/ directories and subdirectories '
-                   'will be kept.')
-def temp_delete(path, warn, force, **kwargs):
-    """Deletes the temporary file or directory.
-
-    Use the command 'perun utils temp delete .' to safely delete all unprotected files in the
-    '.perun/tmp/' directory.
-
-    The command 'perun utils temp delete -f .' can be used to clear the whole '.perun/tmp/'
-    directory including protected files. Protected files are usually more important and should not
-    be deleted without a good reason (such as tmp/ corruption, too many uncleared files etc.)
-
-    However, deleting temp files (protected or not) should not be done when other perun processes
-    are running as it may cause them to crash due to a missing file.
-    """
-    commands.delete_temps(path, not warn, force, **kwargs)
-
-
-@temp_group.command('sync')
-def temp_sync():
-    """Synchronizes the '.perun/tmp/' directory contents with the internal tracking file. This is
-    useful when some files or directories were deleted manually and the resulting inconsistency is
-    causing troubles - however, this should be a very rare condition.
-
-    Invoking the 'temp list' command should also synchronize the internal state automatically.
-    """
-    commands.sync_temps()
-
-
-@utils_group.group('stats')
-def stats_group():
-    """Provides a set of operations for manipulating the stats directory (.perun/stats/) of perun.
-    """
-    pass
-
-
-@stats_group.command('list-files')
-@click.option('--top', '-N', type=int, default=stats.DEFAULT_STATS_LIST_TOP, show_default=True,
-              help='Show only stat files from top N minor versions. Show all results if set to 0. '
-                   'The minor version to start at can be changed using --from-minor.')
-@click.option('--from-minor', '-m', default=None, metavar='<hash>', is_eager=True,
-              callback=cli_helpers.lookup_minor_version_callback,
-              help='Show stat files starting from a certain minor version (default is HEAD).')
-@click.option('--no-minor', '-i', flag_value=True, default=False,
-              help='Do not show the minor version headers in the output.')
-@click.option('--no-file-size', '-f', flag_value=True, default=False,
-              help='Do not show the size of each stat file.')
-@click.option('--no-total-size', '-t', flag_value=True, default=False,
-              help='Do not show the total size of all the stat files combined.')
-@click.option('--sort-by-size', '-s', flag_value=True, default=False,
-              help='Sort the files by size instead of the minor versions order.')
-@click.option('--no-color', '-c', flag_value=True, default=False,
-              help='Disable the output coloring, useful for storing the output to file etc.')
-def stats_list_files(**kwargs):
-    """Show stat files stored in the stats directory (.perun/stats/). This command shows only a
-    limited number of the most recent files by default. This can be, however, changed by the
-    --top and --from-minor options.
-
-    The default output format is 'file size | minor version | file name'.
-    """
-    commands.list_stat_objects('files', **kwargs)
-
-
-@stats_group.command('list-versions')
-@click.option('--top', '-N', type=int, default=stats.DEFAULT_STATS_LIST_TOP, show_default=True,
-              help='Show only top N minor versions. Show all versions if set to 0. '
-                   'The minor version to start at can be changed using --from-minor.')
-@click.option('--from-minor', '-m', default=None, metavar='<hash>', is_eager=True,
-              callback=cli_helpers.lookup_minor_version_callback,
-              help='Show minor versions starting from a certain minor version (default is HEAD).')
-@click.option('--no-dir-size', '-d', flag_value=True, default=False,
-              help='Do not show the size of the version directory.')
-@click.option('--no-file-count', '-f', flag_value=True, default=False,
-              help='Do not show the number of files in each version directory.')
-@click.option('--no-total-size', '-t', flag_value=True, default=False,
-              help='Do not show the total size of all the versions combined.')
-@click.option('--sort-by-size', '-s', flag_value=True, default=False,
-              help='Sort the versions by size instead of their VCS order.')
-@click.option('--no-color', '-c', flag_value=True, default=False,
-              help='Disable the output coloring.')
-def stats_list_versions(**kwargs):
-    """Show minor versions stored as directories in the stats directory (.perun/stats/).
-    This command shows only a limited number of the most recent versions by default. This can be,
-    however, changed by the --top and --from-minor options.
-
-    The default output format is 'directory size | minor version | file count'.
-    """
-    commands.list_stat_objects('versions', **kwargs)
-
-
-@stats_group.group('delete')
-def stats_delete_group():
-    """Allows the deletion of stat files, minor versions or the whole stats directory.
-    """
-    pass
-
-
-@stats_delete_group.command('file')
-@click.argument('name', type=click.Path(), callback=cli_helpers.lookup_stats_file_callback)
-@click.option('--in-minor', '-m', default=None, metavar='<hash>', is_eager=True,
-              callback=cli_helpers.check_stats_minor_callback,
-              help='Delete the stats file in the specified minor version (HEAD if not specified) '
-                   'or across all the minor versions if set to ".".')
-@click.option('--keep-directory', '-k', flag_value=True, default=False,
-              help='Possibly empty directory of minor version will be kept in the file system.')
-def stats_delete_file(**kwargs):
-    """Deletes a stat file in either specific minor version or across all the minor versions in the
-    stats directory.
-    """
-    commands.delete_stats_file(**kwargs)
-
-
-@stats_delete_group.command('minor')
-@click.argument('version', callback=cli_helpers.check_stats_minor_callback)
-@click.option('--keep-directory', '-k', flag_value=True, default=False,
-              help='Resulting empty directory of minor version will be kept in the file system.')
-def stats_delete_minor(version, **kwargs):
-    """Deletes the specified minor version directory in stats with all its content.
-    """
-    commands.delete_stats_minor(version, **kwargs)
-
-
-@stats_delete_group.command('.')
-@click.option('--keep-directory', '-k', flag_value=True, default=False,
-              help='Resulting empty directories of minor versions will be kept in the file system.')
-def stats_delete_all(**kwargs):
-    """Deletes the whole content of the stats directory.
-    """
-    commands.delete_stats_all(**kwargs)
-
-
-@stats_group.command('clean')
-@click.option('--keep-custom', '-c', flag_value=True, default=False,
-              help='The custom stats directories will not be removed.')
-@click.option('--keep-empty', '-e', flag_value=True, default=False,
-              help='The empty version directories will not be removed.')
-def stats_clean(**kwargs):
-    """Cleans the stats directory by synchronizing the internal state, deleting distinguishable
-    custom files and directories (i.e. not all the custom made or manually created files /
-    directories can be identified as custom, e.g. when they comply the correct format etc.)
-    and by removing the empty minor version directories.
-    """
-    commands.clean_stats(**kwargs)
-
-
-@stats_group.command('sync')
-def stats_sync():
-    """ Synchronizes the actual contents of the stats directory with the internal 'index' file.
-    The synchronization should be needed only rarely - mainly in cases when the stats directory
-    has been manually tampered with and some files or directories were created or deleted by a user.
-    """
-    commands.sync_stats()
 
 
 def init_unit_commands(lazy_init=True):
@@ -1361,6 +657,11 @@ def init_unit_commands(lazy_init=True):
 
 # Initialization of other stuff
 init_unit_commands()
+cli.add_command(check_cli.check_group)
+cli.add_command(config_cli.config)
+cli.add_command(run_cli.run)
+cli.add_command(utils_cli.utils_group)
+
 
 if __name__ == "__main__":
     cli()
