+++ perun/cli_groups/__init__.py
@@ -1,0 +1,3 @@
+"""Collection of CLI groups that are not part of the standalone commands"""
+
+__author__ = 'Tomas Fiedor'
